+++ NEWS.md
@@ -4,6 +4,7 @@
 - Ability to retrieve the call wherefrom the issue originates (from the `call` field)
 - `e` and `w` capture the name of the parent function from which they are called
 - Added `raiser` to change the `raise` method.
+- `raise.e` and `raise.w` to define global `raise` methods
 
 # erratum 2.0.0